feat: implement wait_for_server utility for improved server readiness checks
